vtun server host normalization
This commit adds a lowercased normalization to the host struct in vtun. When the server goes to find a host, it will lowercase the hostname and then compare it to both the standard and the normalized hostname. This allows for the server to find the host even if the client sends a hostname with a different case while not affecting how the client operates at all.